Done formatting. This could be considered 1.0
There is a lot of stuff to be cleaned up and optimized, but I feel kind of awful right now. My next focus will be to make it online. Our WotR campaign is on pause for like a month I think, so there's plenty of time for adjustments before it starts being properly used.