changed saveState calls to happen less often
saveState will be called before the user signs out, changes character sheet or clicks on the save button.
Wanted to make it so it also saves when the page unloads (when the user closes or refreshes the widow), but there is an issue with the call (see handleUnload function in useEffect).